+++ src/api/faucet.ts
@@ -6,6 +6,7 @@ import { UserTransactionResponse, WaitForTransactionOptions } from "../types";
 import { AccountAddressInput } from "../core";
 import { AptosConfig } from "./aptosConfig";
 import { waitForIndexer } from "../internal/transaction";
+import { ProcessorType } from "../utils";
 
 /**
  * A class to query all `Faucet` related queries on Aptos.
@@ -33,7 +34,11 @@ export class Faucet {
     const fundTxn = await fundAccount({ aptosConfig: this.config, ...args });
 
     if (args.options?.waitForIndexer !== false) {
-      await waitForIndexer({ aptosConfig: this.config, minimumLedgerVersion: BigInt(fundTxn.version) });
+      await waitForIndexer({
+        aptosConfig: this.config,
+        minimumLedgerVersion: BigInt(fundTxn.version),
+        processorType: ProcessorType.FUNGIBLE_ASSET_PROCESSOR,
+      });
     }
 
     return fundTxn;